Add membership badge
