+++ app/min-side/page.tsx
@@ -1,9 +1,9 @@
 import Title from "@/app/ui/title";
-import UserData from "../components/user-data";
-import { getSessionToken } from "../utils/auth-utils";
-import BlueLink from "../ui/blue-link";
-import BliMedlemButton from "../components/bli-medlem-button";
-import { isUserMember } from "../utils/user-utils";
+import { getSessionToken } from "@/app/utils/auth-utils";
+import BlueLink from "@/app/ui/blue-link";
+import BliMedlemButton from "@/app/components/bli-medlem-button";
+import { isUserMember } from "@/app/utils/user-utils";
+import MembershipBadge from "@/app/components/membership-badge";
 
 async function MyPage() {
   const userId = (await getSessionToken()).payload.userId as number;
@@ -11,8 +11,8 @@ async function MyPage() {
   return (
     <div className="flex flex-col px-4 sm:px-8 gap-8 text-center">
       <Title>Min side</Title>
-      <UserData userId={userId} />
-      {isMember || (
+      <MembershipBadge isMember={isMember} />
+      {!isMember && (
         <div className="flex flex-col gap-4">
           <div className="text-2xl font-semibold text-accent-text">Bli medlem</div>
           <BliMedlemButton paymentType="vipps" />
